Feature un link: continue.

+++ lib/features/setting/application/account_controller.dart
@@ -1,17 +1,32 @@
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_facebook_auth/flutter_facebook_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_sign_in/google_sign_in.dart';
+import 'package:jobs_pot/common/app_keys.dart';
 import 'package:jobs_pot/features/authentication/auth_providers.dart';
 import 'package:jobs_pot/features/authentication/domain/entities/user_entity.dart';
 import 'package:jobs_pot/features/setting/setting_providers.dart';
 import 'package:jobs_pot/resources/i18n/generated/locale_keys.dart';
 import 'package:jobs_pot/system/system_providers.dart';
 import 'package:jobs_pot/utils/utils.dart';
+import 'package:reactive_forms/reactive_forms.dart';
 
-class AccountController extends StateNotifier {
+class AccountController extends StateNotifier<String?> {
   AccountController(this._ref) : super(null);
   final Ref _ref;
+
+  final _passwordForm = FormGroup({
+    ValidationKeys.password: FormControl<String>(
+      value: '',
+      validators: [
+        Validators.required,
+      ],
+      touched: true,
+    ),
+  });
+
+  FormGroup getPasswordForm() => _passwordForm;
 
   final GoogleSignIn _googleSignIn = GoogleSignIn();
 
@@ -19,8 +34,67 @@ class AccountController extends StateNotifier {
     await _googleSignIn.disconnect();
   }
 
-  Future<void> facebookLink() async {
+  void setProvider(String providerId) {
+    state = providerId;
+  }
+
+  void showLoading() {
     _ref.read(systemControllerProvider.notifier).showLoading();
+  }
+
+  void hideLoading() {
+    _ref.read(systemControllerProvider.notifier).hideLoading();
+  }
+
+  void showToastGeneralError() {
+    _ref.read(systemControllerProvider.notifier).showToastGeneralError();
+  }
+
+  String getPassword() {
+    return _passwordForm.controls[ValidationKeys.password]?.value.toString() ??
+        "";
+  }
+
+  void passwordVerify(Function affterVerifyFun, BuildContext context) async {
+    UserEntity? userData = _ref.read(authControllerProvider);
+
+    final isValid = _passwordForm.valid;
+
+    if (isValid) {
+      String password = getPassword();
+      String? email = userData?.email;
+
+      try {
+        if (email != null) {
+          await FirebaseAuth.instance
+              .signInWithEmailAndPassword(email: email, password: password);
+
+          _ref.read(systemControllerProvider.notifier).showToastMessage(
+              Utils.getLocaleMessage(LocaleKeys.settingAccountVerifyMsg));
+
+          affterVerifyFun();
+        }
+      } on FirebaseAuthException catch (_) {
+        _ref.read(systemControllerProvider.notifier).showToastMessage(
+              Utils.getLocaleMessage(
+                LocaleKeys.errorPassword2,
+              ),
+            );
+      }
+    } else {
+      _passwordForm.controls.forEach((key, value) {
+        if (value.invalid) {
+          value.setErrors({
+            ValidationKeys.required: LocaleKeys.authenticationInputRequired
+          });
+        }
+      });
+    }
+  }
+
+  Future<void> facebookLink() async {
+    showLoading();
+    await _ref.read(authControllerProvider.notifier).reloadFirebaseUser();
 
     UserEntity? userData = _ref.read(authControllerProvider);
 
@@ -37,28 +111,30 @@ class AccountController extends StateNotifier {
       );
       final providerEmail = providerData['email'];
       if (userData.email != providerEmail) {
-        _ref.read(systemControllerProvider.notifier).hideLoading();
+        hideLoading();
         _ref.read(systemControllerProvider.notifier).showToastMessage(
             Utils.getLocaleMessage(LocaleKeys.settingAccountLinkAccountError));
       } else {
         await linkAccount(facebookAuthCredential);
-        _ref.read(systemControllerProvider.notifier).hideLoading();
+        hideLoading();
       }
     } else {
-      _ref.read(systemControllerProvider.notifier).hideLoading();
+      hideLoading();
     }
     await FacebookAuth.instance.logOut();
   }
 
   Future<void> googleLink() async {
-    _ref.read(systemControllerProvider.notifier).showLoading();
+    showLoading();
+    await _ref.read(authControllerProvider.notifier).reloadFirebaseUser();
+
     UserEntity? userData = _ref.read(authControllerProvider);
 
     final GoogleSignInAccount? googleUser = await _googleSignIn.signIn();
 
     if (googleUser != null && userData != null) {
       if (userData.email != googleUser.email) {
-        _ref.read(systemControllerProvider.notifier).hideLoading();
+        hideLoading();
 
         _ref.read(systemControllerProvider.notifier).showToastMessage(
             Utils.getLocaleMessage(LocaleKeys.settingAccountLinkAccountError));
@@ -71,27 +147,36 @@ class AccountController extends StateNotifier {
           idToken: googleAuth.idToken,
         );
         await linkAccount(credential);
-        _ref.read(systemControllerProvider.notifier).hideLoading();
+        hideLoading();
       }
     } else {
-      _ref.read(systemControllerProvider.notifier).hideLoading();
+      hideLoading();
     }
     await disconnect();
   }
 
-  Future<void> unLink(String providerId) async {
-    _ref.read(systemControllerProvider.notifier).showLoading();
+  Future<void> unLink() async {
+    showLoading();
     try {
-      await FirebaseAuth.instance.currentUser?.unlink(providerId);
-      await synUnLinkWithServer(providerId);
-      _ref.read(systemControllerProvider.notifier).hideLoading();
+      String? providerId = state;
+      if (providerId != null) {
+        await FirebaseAuth.instance.currentUser?.unlink(providerId);
+        await synUnLinkWithServer(providerId);
+        state = null;
+        hideLoading();
+      } else {
+        _ref.read(systemControllerProvider.notifier).showToastGeneralError();
+
+        hideLoading();
+      }
     } on FirebaseAuthException catch (e) {
+      state = null;
       _ref.read(systemControllerProvider.notifier).handlerFirebaseError(e.code);
     }
   }
 
   Future<void> linkAccount(OAuthCredential credential) async {
-    _ref.read(systemControllerProvider.notifier).showLoading();
+    showLoading();
 
     try {
       UserCredential? userCredential = await FirebaseAuth.instance.currentUser
@@ -110,7 +195,7 @@ class AccountController extends StateNotifier {
           }
         }
       } else {
-        _ref.read(systemControllerProvider.notifier).showToastGeneralError();
+        showToastGeneralError();
       }
     } on FirebaseAuthException catch (e) {
       _ref.read(systemControllerProvider.notifier).handlerFirebaseError(e.code);
@@ -128,7 +213,7 @@ class AccountController extends StateNotifier {
         _ref.read(authControllerProvider.notifier).setDataUser(r.results);
       });
     } else {
-      _ref.read(systemControllerProvider.notifier).showToastGeneralError();
+      showToastGeneralError();
     }
   }
 
@@ -144,7 +229,7 @@ class AccountController extends StateNotifier {
         _ref.read(systemControllerProvider.notifier).showToastMessage(r.msg);
       });
     } else {
-      _ref.read(systemControllerProvider.notifier).showToastGeneralError();
+      showToastGeneralError();
     }
   }
 }
